third_party/WebKit/LayoutTests/shadow-dom/imperative-api.html

<!DOCTYPE html>
<!--
Tests for Imperative Shadow DOM Distribution API.
See https://crbug.com/869308
-->
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="host">
  <div id="child1"></div>
  <div id="child2"></div>
  <div id="child3"></div>
</div>
<div id="host1">
</div>
</div>
<div id="host2">
</div>
</div>
<div id="host3">
</div>


<script>
const host = document.querySelector('#host');
const host1 = document.querySelector('#host1');
const host2 = document.querySelector('#host2');
const host3 = document.querySelector('#host3');
const child1 = document.querySelector('#child1');
const child2 = document.querySelector('#child2');
const child3 = document.querySelector('#child3');
const shadow_root = host.attachShadow({ mode: 'open', slotting: 'manual' });
const slot1 = document.createElement('slot');
const slot2 = document.createElement('slot');
shadow_root.appendChild(slot1);
shadow_root.appendChild(slot2);

test(() => {
  assert_not_equals(host1.attachShadow({ mode: 'open', slotting: 'manual' }),
                    null, 'slotting manual should work');
  assert_not_equals(host2.attachShadow({ mode: 'open', slotting: 'auto' }),
                    null, 'slotting auto should work');
  assert_throws(new TypeError(), () => {
    host3.attachShadow({ mode: 'open', slotting: 'exceptional' })},
                'others should throw exception');
}, 'attachShadow can take slotting parameter');

test(() => {
  assert_array_equals(slot1.assignedElements(), []);
  assert_equals(child1.assignedSlot, null);

  slot1.assign([child1]);

  assert_array_equals(slot1.assignedNodes(), [child1]);
  assert_equals(child1.assignedSlot, slot1);

  slot2.assign([child1]);
  assert_array_equals(slot1.assignedNodes(), []);
  assert_equals(child1.assignedSlot, slot2);

  slot1.assign([child2,child3]);
  assert_array_equals(slot1.assignedNodes(), [child2,child3]);
}, 'assignedNodes can be used in manual slotting');
</script>
